requirements.txt
piper-tts
pynput
pyperclip